Split AgentBrowser into multiple components
