--- front/components/assistant/AgentBrowser.tsx
@@ -1,726 +1,0 @@
-import { CreateDropdown } from "@app/components/assistant/CreateDropdown";
-import { AgentDetailsDropdownMenu } from "@app/components/assistant/details/AgentDetailsButtonBar";
-import { AgentDetailsSheet } from "@app/components/assistant/details/AgentDetailsSheet";
-import { rankAgentsByPopularity } from "@app/components/assistant/helpers/agents";
-import { ManageDropdownMenu } from "@app/components/assistant/ManageDropdownMenu";
-import { useWelcomeTourGuide } from "@app/components/assistant/WelcomeTourGuideProvider";
-import { useHashParam } from "@app/hooks/useHashParams";
-import { usePersistedAgentBrowserSelection } from "@app/hooks/usePersistedAgentBrowserSelection";
-import { useFeatureFlags } from "@app/lib/auth/AuthContext";
-import { useClientType } from "@app/lib/context/clientType";
-import { useAppRouter } from "@app/lib/platform";
-import { useIsMobile } from "@app/lib/swr/useIsMobile";
-import { TRACKING_AREAS, withTracking } from "@app/lib/tracking";
-import {
-  compareForFuzzySort,
-  getAgentSearchString,
-  subFilter,
-  tagsSorter,
-} from "@app/lib/utils";
-import { getAgentBuilderRoute, setQueryParam } from "@app/lib/utils/router";
-import type { LightAgentConfigurationType } from "@app/types/assistant/agent";
-import { compareAgentsForSort } from "@app/types/assistant/assistant";
-import type { TagType } from "@app/types/tag";
-import type { UserType, WorkspaceType } from "@app/types/user";
-import { isBuilder } from "@app/types/user";
-import {
-  AssistantCard,
-  AssistantCardMore,
-  Avatar,
-  Button,
-  CardGrid,
-  Chip,
-  ContactsRobotIcon,
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuTrigger,
-  MoreIcon,
-  ScrollArea,
-  ScrollBar,
-  SearchDropdownMenu,
-  Spinner,
-  Tabs,
-  TabsList,
-  TabsTrigger,
-} from "@dust-tt/sparkle";
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
-import { useInView } from "react-intersection-observer";
-
-function isValidTab(tab: string, visibleTabs: TabId[]): tab is TabId {
-  return visibleTabs.includes(tab as TabId);
-}
-
-const AGENTS_TABS = [
-  { label: "Favorites", id: "favorites" },
-  { label: "All agents", id: "all" },
-  { label: "Editable by me", id: "editable_by_me" },
-] as const;
-
-type TabId = (typeof AGENTS_TABS)[number]["id"];
-
-const MOST_POPULAR_TAG: TagType = {
-  sId: "--most_popular--",
-  name: "Most popular",
-  kind: "protected",
-};
-
-const ALL_TAG: TagType = {
-  sId: "--all--",
-  name: "All",
-  kind: "protected",
-};
-
-const OTHERS_TAG: TagType = {
-  sId: "--others--",
-  name: "Others",
-  kind: "protected",
-};
-
-type AgentGridProps = {
-  agentConfigurations: LightAgentConfigurationType[];
-  handleAssistantClick: (agent: LightAgentConfigurationType) => void;
-  handleMoreClick: (agentId: string) => void;
-  owner: WorkspaceType;
-};
-
-export const AgentGrid = ({
-  agentConfigurations,
-  handleAssistantClick,
-  handleMoreClick,
-  owner,
-}: AgentGridProps) => {
-  // Context menu state
-  const [contextMenuAgent, setContextMenuAgent] =
-    useState<LightAgentConfigurationType | null>(null);
-  const [contextMenuPosition, setContextMenuPosition] = useState<{
-    x: number;
-    y: number;
-  } | null>(null);
-
-  const clientType = useClientType();
-  const isMobile = useIsMobile();
-
-  // Handle "infinite" scroll
-  // We only start with 9 items shown (no need more on mobile) and load more until we fill the parent container.
-  // We use an intersection observer to detect when the bottom of the list is visible and load more items.
-  // That way, the list starts lightweight and only show more items when needed.
-  const ITEMS_PER_PAGE = 9; // Should be a multiple of 3
-
-  const [itemsPage, setItemsPage] = useState(0);
-
-  // biome-ignore lint/correctness/useExhaustiveDependencies: ignored using `--suppress`
-  const nextPage = useCallback(() => {
-    setItemsPage(itemsPage + 1);
-  }, [setItemsPage, itemsPage]);
-
-  const previousEntry = useRef<IntersectionObserverEntry | undefined>(
-    undefined
-  );
-
-  const { ref, inView, entry } = useInView({
-    threshold: 0,
-  });
-
-  useEffect(() => {
-    if (
-      // The observer is in view.
-      inView &&
-      // We have more items to show.
-      agentConfigurations.length > itemsPage * ITEMS_PER_PAGE &&
-      // The entry is different from the previous one to avoid multiple calls for the same intersection.
-      entry != previousEntry.current
-    ) {
-      previousEntry.current = entry;
-      nextPage();
-    }
-  }, [inView, nextPage, entry, agentConfigurations.length, itemsPage]);
-
-  const slicedAgentConfigurations = agentConfigurations.slice(
-    0,
-    (itemsPage + 1) * ITEMS_PER_PAGE
-  );
-
-  const isMobileOrExtension = clientType === "extension" || isMobile;
-
-  return (
-    <>
-      <CardGrid>
-        {slicedAgentConfigurations.map((agent, index) => {
-          const isLastItem = index === slicedAgentConfigurations.length - 1;
-          return (
-            <AssistantCard
-              // Force a re-render of the last item to trigger the intersection observer
-              key={isLastItem ? `${agent.sId}-${itemsPage}` : agent.sId}
-              ref={isLastItem ? ref : undefined}
-              title={agent.name}
-              pictureUrl={agent.pictureUrl}
-              subtitle={agent.lastAuthors?.join(", ") ?? ""}
-              description={agent.description}
-              onClick={() => handleAssistantClick(agent)}
-              onContextMenu={(e) => {
-                e.preventDefault();
-                setContextMenuAgent(agent);
-                setContextMenuPosition({ x: e.clientX, y: e.clientY });
-              }}
-              iconSize={isMobileOrExtension ? "sm" : "md"}
-              action={
-                isMobileOrExtension ? undefined : (
-                  <AssistantCardMore
-                    onClick={(e: React.MouseEvent<HTMLButtonElement>) => {
-                      e.stopPropagation();
-                      handleMoreClick(agent.sId);
-                    }}
-                  />
-                )
-              }
-            />
-          );
-        })}
-      </CardGrid>
-      <AgentDetailsDropdownMenu
-        agentConfiguration={contextMenuAgent ?? undefined}
-        owner={owner}
-        onClose={() => {
-          setContextMenuPosition(null);
-        }}
-        showEditOption={true}
-        contextMenuPosition={contextMenuPosition ?? undefined}
-      />
-    </>
-  );
-};
-
-interface AssistantBrowserProps {
-  owner: WorkspaceType;
-  agentConfigurations: LightAgentConfigurationType[];
-  isLoading: boolean;
-  handleAgentClick: (agent: LightAgentConfigurationType) => void;
-  user: UserType;
-}
-
-export function AgentBrowser({
-  owner,
-  agentConfigurations,
-  isLoading,
-  handleAgentClick,
-  user,
-}: AssistantBrowserProps) {
-  const [assistantSearch, setAssistantSearch] = useState<string>("");
-  const [selectedTab, setSelectedTab] = useHashParam(
-    "selectedTab",
-    "favorites"
-  );
-  const [displayedAssistantId, setDisplayedAssistantId] = useState<
-    string | null
-  >(null);
-
-  const router = useAppRouter();
-  const clientType = useClientType();
-  const { createAgentButtonRef } = useWelcomeTourGuide();
-  const [selectedTag, setSelectedTag] = useState<string | null>(null);
-  const [sortType, setSortType] = useState<
-    "popularity" | "alphabetical" | "updated"
-  >("popularity");
-
-  const isMobile = useIsMobile();
-
-  const { featureFlags } = useFeatureFlags();
-
-  const { selectedTagId: persistedSelectedTagId, setSelectedTagId } =
-    usePersistedAgentBrowserSelection(owner.sId);
-
-  const isRestrictedFromAgentCreation =
-    featureFlags.includes("disallow_agent_creation_to_users") &&
-    !isBuilder(owner);
-
-  const sortAgents = useCallback(
-    (a: LightAgentConfigurationType, b: LightAgentConfigurationType) => {
-      if (sortType === "popularity") {
-        return (
-          (b.usage?.messageCount ?? 0) - (a.usage?.messageCount ?? 0) ||
-          a.name.toLowerCase().localeCompare(b.name.toLowerCase())
-        );
-      }
-      if (sortType === "updated") {
-        return (
-          new Date(b.versionCreatedAt ?? 0).getTime() -
-            new Date(a.versionCreatedAt ?? 0).getTime() ||
-          a.name.toLowerCase().localeCompare(b.name.toLowerCase())
-        );
-      }
-      return a.name.toLowerCase().localeCompare(b.name.toLowerCase());
-    },
-    [sortType]
-  );
-
-  const agentsByTab = useMemo(() => {
-    const allAgents: LightAgentConfigurationType[] = agentConfigurations
-      .filter((a) => a.status === "active")
-      .sort(sortAgents);
-
-    return {
-      // do not show the "all" tab while still loading all agents
-      all: allAgents,
-      favorites: allAgents.filter((a) => a.userFavorite),
-      editable_by_me: allAgents.filter((a) => a.canEdit),
-      most_popular: rankAgentsByPopularity(allAgents),
-    };
-  }, [agentConfigurations, sortAgents]);
-
-  const { filteredAgents, filteredTags, uniqueTags, noTagsDefined } =
-    useMemo(() => {
-      const tags = agentConfigurations.flatMap((a) => a.tags);
-      // Remove duplicate tags by unique sId
-      const uniqueTags = Array.from(
-        new Map(tags.map((tag) => [tag.sId, tag])).values()
-      ).sort(tagsSorter);
-
-      uniqueTags.unshift(ALL_TAG);
-      uniqueTags.unshift(MOST_POPULAR_TAG);
-
-      // Always append others at the end
-      uniqueTags.push(OTHERS_TAG);
-
-      if (assistantSearch.trim() === "") {
-        return {
-          filteredAgents: [],
-          filteredTags: [],
-          uniqueTags,
-          noTagsDefined: uniqueTags.length === 2, // Only most popular and others
-        };
-      }
-      const search = assistantSearch.toLowerCase().trim().replace(/^@/, "");
-
-      const filteredAgents: LightAgentConfigurationType[] = agentConfigurations
-        .filter(
-          (a) =>
-            a.status === "active" &&
-            // Filters on search query
-            subFilter(search, getAgentSearchString(a))
-        )
-
-        .sort((a, b) => {
-          return (
-            compareForFuzzySort(
-              assistantSearch.toLowerCase(),
-              getAgentSearchString(a),
-              getAgentSearchString(b)
-            ) ||
-            (b.usage?.messageCount ?? 0) - (a.usage?.messageCount ?? 0) ||
-            compareAgentsForSort(a, b)
-          );
-        });
-
-      const filteredTags =
-        selectedTag === ALL_TAG.sId
-          ? uniqueTags
-          : uniqueTags.filter((t) => subFilter(search, t.name.toLowerCase()));
-
-      return {
-        filteredAgents,
-        filteredTags,
-        uniqueTags,
-        noTagsDefined: uniqueTags.length === 2, // Only most popular and others
-      };
-    }, [agentConfigurations, assistantSearch, selectedTag]);
-
-  // check the query string for the tab to show, the query param to look for is called "selectedTab"
-  // if it's not found, show the first tab with agents
-  const viewTab = useMemo(() => {
-    const enabledTabs = AGENTS_TABS.filter(
-      (tab) => agentsByTab[tab.id].length > 0
-    );
-    return selectedTab &&
-      isValidTab(
-        selectedTab,
-        enabledTabs.map((tab) => tab.id)
-      )
-      ? selectedTab
-      : enabledTabs[0]?.id;
-  }, [selectedTab, agentsByTab]);
-
-  // Initialize `selectedTag` from persisted selection (or default to Most popular).
-  // biome-ignore lint/correctness/useExhaustiveDependencies: ignored using `--suppress`
-  useEffect(() => {
-    if (noTagsDefined || selectedTag) {
-      return;
-    }
-
-    const validTagIds = new Set(uniqueTags.map((t) => t.sId));
-    const persistedValid =
-      persistedSelectedTagId && validTagIds.has(persistedSelectedTagId);
-
-    if (persistedValid) {
-      setSelectedTag(persistedSelectedTagId);
-    } else {
-      setSelectedTag(MOST_POPULAR_TAG.sId);
-    }
-  }, [
-    noTagsDefined,
-    persistedSelectedTagId,
-    uniqueTags,
-    selectedTag,
-    setSelectedTag,
-  ]);
-
-  // Persist selectedTag when they change (and tags exist).
-  useEffect(() => {
-    if (noTagsDefined) {
-      return;
-    }
-
-    if (persistedSelectedTagId !== selectedTag) {
-      setSelectedTagId(selectedTag);
-    }
-  }, [selectedTag, noTagsDefined, persistedSelectedTagId, setSelectedTagId]);
-
-  const sortTypeLabel = useMemo(() => {
-    switch (sortType) {
-      case "popularity":
-        return "By popularity";
-      case "alphabetical":
-        return "Alphabetical";
-      case "updated":
-        return "Recently updated";
-    }
-  }, [sortType]);
-
-  const isMobileOrExtension = clientType === "extension" || isMobile;
-
-  return (
-    <>
-      {/* Search bar */}
-      {!isMobileOrExtension && (
-        <div
-          id="search-container"
-          className="mb-2 flex w-full flex-row items-center justify-center gap-2 align-middle"
-        >
-          <SearchDropdownMenu
-            searchInputValue={assistantSearch}
-            setSearchInputValue={setAssistantSearch}
-          >
-            {filteredTags.length > 0 || filteredAgents.length > 0 ? (
-              <>
-                {filteredTags.length > 0 && <DropdownMenuLabel label="Tags" />}
-                {filteredTags.map((tag) => (
-                  <DropdownMenuItem
-                    key={tag.sId}
-                    onClick={() => {
-                      setSelectedTab("all");
-                      setAssistantSearch("");
-                      setSelectedTag(tag.sId);
-                      setTimeout(() => {
-                        const element = document.getElementById(
-                          `anchor-${tag.sId}`
-                        );
-                        if (element) {
-                          element.scrollIntoView({
-                            behavior: "smooth",
-                            block: "start",
-                            inline: "nearest",
-                          });
-                        }
-                      }, 300); // Need to wait for the dropdown to close before scrolling
-                    }}
-                  >
-                    <Chip label={tag.name} color="golden" size="xs" />
-                  </DropdownMenuItem>
-                ))}
-                {filteredAgents.length > 0 && (
-                  <DropdownMenuLabel label="Agents" />
-                )}
-                {filteredAgents.map((agent) => (
-                  <DropdownMenuItem
-                    key={agent.sId}
-                    onClick={() => {
-                      handleAgentClick(agent);
-                      setAssistantSearch("");
-                    }}
-                    truncateText
-                    label={agent.name}
-                    description={agent.description}
-                    icon={() => <Avatar size="sm" visual={agent.pictureUrl} />}
-                    endComponent={
-                      <Button
-                        variant="ghost"
-                        size="xs"
-                        icon={MoreIcon}
-                        onClick={(e: React.MouseEvent<HTMLButtonElement>) => {
-                          e.stopPropagation();
-                          setQueryParam(router, "agentDetails", agent.sId);
-                        }}
-                      />
-                    }
-                  />
-                ))}
-              </>
-            ) : isLoading ? (
-              <div className="flex justify-center py-8">
-                <Spinner size="md" />
-              </div>
-            ) : (
-              <div className="p-2 text-sm text-gray-500">No results found</div>
-            )}
-          </SearchDropdownMenu>
-
-          {clientType === "web" && (
-            <div className="hidden sm:block">
-              <div className="flex gap-2">
-                {!isRestrictedFromAgentCreation && (
-                  <div ref={createAgentButtonRef}>
-                    <CreateDropdown owner={owner} dataGtmLocation="homepage" />
-                  </div>
-                )}
-
-                {isBuilder(owner) ? (
-                  <ManageDropdownMenu owner={owner} />
-                ) : (
-                  <Button
-                    href={getAgentBuilderRoute(owner.sId, "manage")}
-                    variant="primary"
-                    icon={ContactsRobotIcon}
-                    label="Manage agents"
-                    data-gtm-label="assistantManagementButton"
-                    data-gtm-location="homepage"
-                    size="sm"
-                    onClick={withTracking(
-                      TRACKING_AREAS.BUILDER,
-                      "manage_agents"
-                    )}
-                  />
-                )}
-              </div>
-            </div>
-          )}
-        </div>
-      )}
-
-      <AgentDetailsSheet
-        owner={owner}
-        user={user}
-        agentId={displayedAssistantId}
-        onClose={() => setDisplayedAssistantId(null)}
-      />
-
-      {/* Agent tabs */}
-      {isMobileOrExtension ? (
-        <div className="w-full flex flex-row gap-2">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button
-                isSelect
-                variant="outline"
-                label={
-                  viewTab
-                    ? AGENTS_TABS.find((tab) => tab.id === viewTab)?.label
-                    : "Select view"
-                }
-                size="sm"
-              />
-            </DropdownMenuTrigger>
-            <DropdownMenuContent>
-              {AGENTS_TABS.map((tab) => (
-                <DropdownMenuItem
-                  key={tab.id}
-                  disabled={agentsByTab[tab.id].length === 0}
-                  onClick={() => setSelectedTab(tab.id)}
-                  label={tab.label}
-                />
-              ))}
-            </DropdownMenuContent>
-          </DropdownMenu>
-          <SearchDropdownMenu
-            searchInputValue={assistantSearch}
-            setSearchInputValue={setAssistantSearch}
-          >
-            {filteredTags.length > 0 || filteredAgents.length > 0 ? (
-              <>
-                {filteredTags.length > 0 && <DropdownMenuLabel label="Tags" />}
-                {filteredTags.map((tag) => (
-                  <DropdownMenuItem
-                    key={tag.sId}
-                    onClick={() => {
-                      setSelectedTab("all");
-                      setAssistantSearch("");
-                      setSelectedTag(tag.sId);
-                      setTimeout(() => {
-                        const element = document.getElementById(
-                          `anchor-${tag.sId}`
-                        );
-                        if (element) {
-                          element.scrollIntoView({
-                            behavior: "smooth",
-                            block: "start",
-                            inline: "nearest",
-                          });
-                        }
-                      }, 300); // Need to wait for the dropdown to close before scrolling
-                    }}
-                  >
-                    <Chip label={tag.name} color="golden" size="xs" />
-                  </DropdownMenuItem>
-                ))}
-                {filteredAgents.length > 0 && (
-                  <DropdownMenuLabel label="Agents" />
-                )}
-                {filteredAgents.map((agent) => (
-                  <DropdownMenuItem
-                    key={agent.sId}
-                    onClick={() => {
-                      handleAgentClick(agent);
-                      setAssistantSearch("");
-                    }}
-                    truncateText
-                    label={agent.name}
-                    description={agent.description}
-                    icon={() => <Avatar size="sm" visual={agent.pictureUrl} />}
-                  />
-                ))}
-              </>
-            ) : isLoading ? (
-              <div className="flex justify-center py-8">
-                <Spinner size="md" />
-              </div>
-            ) : (
-              <div className="p-2 text-sm text-gray-500">No results found</div>
-            )}
-          </SearchDropdownMenu>
-        </div>
-      ) : (
-        <div className="w-full">
-          <ScrollArea aria-orientation="horizontal">
-            <Tabs value={viewTab} onValueChange={setSelectedTab}>
-              <TabsList>
-                {AGENTS_TABS.map((tab) => (
-                  <TabsTrigger
-                    disabled={agentsByTab[tab.id].length === 0}
-                    key={tab.id}
-                    value={tab.id}
-                    label={tab.label}
-                  />
-                ))}
-                <div className="ml-auto"></div>
-                <DropdownMenu>
-                  <DropdownMenuTrigger asChild>
-                    <Button
-                      isSelect
-                      variant="outline"
-                      label={sortTypeLabel}
-                      size="sm"
-                    />
-                  </DropdownMenuTrigger>
-
-                  <DropdownMenuContent>
-                    <DropdownMenuItem
-                      label="By popularity"
-                      onClick={() => setSortType("popularity")}
-                    />
-                    <DropdownMenuItem
-                      label="Alphabetical"
-                      onClick={() => setSortType("alphabetical")}
-                    />
-                    <DropdownMenuItem
-                      label="Recently updated"
-                      onClick={() => setSortType("updated")}
-                    />
-                  </DropdownMenuContent>
-                </DropdownMenu>
-              </TabsList>
-            </Tabs>
-            <ScrollBar orientation="horizontal" className="hidden" />
-          </ScrollArea>
-        </div>
-      )}
-
-      {viewTab === "all" ? (
-        <>
-          <div className="mb-2 flex flex-wrap items-center gap-2">
-            {noTagsDefined ? null : (
-              <>
-                {uniqueTags.map((tag) => (
-                  <Button
-                    size="xs"
-                    variant={selectedTag === tag.sId ? "primary" : "outline"}
-                    key={tag.sId}
-                    label={tag.name}
-                    onClick={() => {
-                      setSelectedTag(tag.sId);
-                    }}
-                  />
-                ))}
-              </>
-            )}
-          </div>
-
-          <div className="flex flex-col gap-4">
-            {uniqueTags
-              .filter(
-                (t) =>
-                  // User picked specific tag(s).
-                  selectedTag === t.sId ||
-                  // No tags are defined, show most popular & others.
-                  noTagsDefined
-              )
-              .map((tag) => (
-                <React.Fragment key={tag.sId}>
-                  <a id={`anchor-${tag.sId}`} />
-                  {!isMobileOrExtension && (
-                    <span className="heading-base">{tag.name}</span>
-                  )}
-                  <AgentGrid
-                    agentConfigurations={agentsByTab.all.filter((a) => {
-                      return (
-                        // One of the tags is the selected tag.
-                        a.tags.some((t) => t.sId === tag.sId) ||
-                        // Selected tag is others, and the agent has no tags.
-                        (tag.sId === OTHERS_TAG.sId &&
-                          a.tags.length === 0 &&
-                          // Exclude agents that are in the most popular list.
-                          !agentsByTab.most_popular.some(
-                            (a_popular) => a_popular.sId === a.sId
-                          )) ||
-                        // Selected tag is most popular, and the agent is in the most popular list.
-                        (tag.sId === MOST_POPULAR_TAG.sId &&
-                          agentsByTab.most_popular.some(
-                            (a_popular) => a_popular.sId === a.sId
-                          )) ||
-                        (tag.sId === ALL_TAG.sId &&
-                          agentsByTab.all.some(
-                            (a_popular) => a_popular.sId === a.sId
-                          ))
-                      );
-                    })}
-                    handleAssistantClick={handleAgentClick}
-                    handleMoreClick={setDisplayedAssistantId}
-                    owner={owner}
-                  />
-                </React.Fragment>
-              ))}
-          </div>
-        </>
-      ) : (
-        viewTab && (
-          <AgentGrid
-            agentConfigurations={agentsByTab[viewTab]}
-            handleAssistantClick={handleAgentClick}
-            handleMoreClick={setDisplayedAssistantId}
-            owner={owner}
-          />
-        )
-      )}
-      {isLoading && (
-        <div className="flex justify-center py-8">
-          <Spinner size="lg" />
-        </div>
-      )}
-    </>
-  );
-}
